added a Message Bus agent base class
